Write counter module to count how many times a request has received
Write a counter module which logs the number of requests per URL to a file called ".logs". At the time of writing, this file has to be created manually, otherwise things break down. This module expose methods for each URL to keep track of matching requests. Since logging is done by reading/writing to the file, it might effect the performance; Thus it is advised to use this this module at the end of each router callback.